Refactored client API and fixed some minor bugs

+++ src/Clustron.Abstractions/ClusterMetricsSnapshot.cs
@@ -4,7 +4,7 @@
 // included in the LICENSE file in the root of this repository.
 //
 // Production use is not permitted without a commercial license from the Licensor.
-// To obtain a license for production, please contact: heartbeats.zero@gmail.com
+// To obtain a license for production, please contact: support@clustron.io
 
 using System;
 using System.Collections.Generic;
@@ -21,7 +21,7 @@ public class ClusterMetricsSnapshot
 
     public int ActiveConnections { get; set; }
 
-    public Dictionary<string, int> Totals { get; set; } = new();
+    public Dictionary<string, long> Totals { get; set; } = new();
     public Dictionary<string, int[]> PerSecondRates { get; set; } = new();
 
     // Optional single-value metrics (e.g., for gauges)
